Move testing_utils.go to testing_utils_test.go for SCEP tests
