+++ ee/server/service/scep/sceptest/sceptest.go
@@ -1,4 +1,11 @@
-package scep
+// Package sceptest provides test helpers (SCEP test server, NDES admin
+// page emulator, dynamic challenge server) for tests that exercise the SCEP
+// proxy.
+//
+// It imports the "testing" package and must therefore only ever be imported
+// from test code; importing it from production code would pull "testing"
+// into the resulting binary.
+package sceptest
 
 import (
 	"crypto/x509"
@@ -27,8 +34,17 @@ var caKey []byte
 //go:embed testdata/testca/ca.pem
 var caPem []byte
 
-// NewTestSCEPServer creates a new SCEP server for testing purposes. The depotPath should be the
-// relative path to the directory where the test CA files are stored (e.g., "./testdata/testca")
+//go:embed testdata/mscep_admin_cache_full.html
+var mscepAdminCacheFull []byte
+
+//go:embed testdata/mscep_admin_insufficient_permissions.html
+var mscepAdminInsufficientPermissions []byte
+
+//go:embed testdata/mscep_admin_password.html
+var mscepAdminPassword []byte
+
+// NewTestSCEPServer creates a new SCEP server for testing purposes, backed by
+// an in-temp-dir CA built from embedded test certs.
 func NewTestSCEPServer(t *testing.T) *httptest.Server {
 	t.Helper()
 
@@ -75,25 +91,21 @@ func (d *noopDepot) Put(_ string, _ *x509.Certificate) error {
 	return nil
 }
 
-//go:embed testdata/mscep_admin_cache_full.html
-var mscepAdminCacheFull []byte
-
-//go:embed testdata/mscep_admin_insufficient_permissions.html
-var mscepAdminInsufficientPermissions []byte
-
-//go:embed testdata/mscep_admin_password.html
-var mscepAdminPassword []byte
-
-func NewTestNDESAdminServer(t *testing.T, responseTemplate string, responseStatus int) *httptest.Server {
+// NewTestNDESAdminServer creates an httptest.Server that emulates the NDES
+// admin page. responseTemplate selects which canned response to serve:
+// "mscep_admin_cache_full", "mscep_admin_insufficient_permissions",
+// "mscep_admin_password", or anything else for an empty body.
+func NewTestNDESAdminServer(t *testing.T, responseTemplate string, _ int) *httptest.Server {
 	t.Helper()
 
 	var returnPage func() []byte
 	returnStatus := http.StatusOK
-	ndesAdminServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+	ndesAdminServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 		w.WriteHeader(returnStatus)
 		if returnStatus == http.StatusOK {
-			_, err := w.Write(returnPage())
-			require.NoError(t, err)
+			if _, err := w.Write(returnPage()); err != nil {
+				t.Errorf("write NDES admin response: %v", err)
+			}
 		}
 	}))
 	t.Cleanup(ndesAdminServer.Close)
@@ -111,17 +123,14 @@ func NewTestNDESAdminServer(t *testing.T, responseTemplate string, responseStatu
 
 	switch responseTemplate {
 	case "mscep_admin_cache_full":
-		// Catch ths issue when NDES password cache is full
 		returnPage = func() []byte {
 			return convertHTML(mscepAdminCacheFull)
 		}
 	case "mscep_admin_insufficient_permissions":
-		// Catch this issue when account has insufficient permissions
 		returnPage = func() []byte {
 			return convertHTML(mscepAdminInsufficientPermissions)
 		}
 	case "mscep_admin_password":
-		// All good, NDES admin page loads correctly
 		returnPage = func() []byte {
 			return convertHTML(mscepAdminPassword)
 		}
@@ -134,22 +143,27 @@ func NewTestNDESAdminServer(t *testing.T, responseTemplate string, responseStatu
 	return ndesAdminServer
 }
 
+// NewTestDynamicChallengeServer creates an httptest.Server that emulates a
+// dynamic SCEP challenge endpoint, always returning the string
+// "dynamic challenge".
 func NewTestDynamicChallengeServer(t *testing.T) *httptest.Server {
 	t.Helper()
 
 	dynamicChallengeServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(http.StatusOK)
 		fmt.Println(r.URL.Path)
-		_, err := w.Write([]byte("dynamic challenge"))
-		require.NoError(t, err)
+		if _, err := w.Write([]byte("dynamic challenge")); err != nil {
+			t.Errorf("write dynamic challenge response: %v", err)
+		}
 	}))
 	t.Cleanup(dynamicChallengeServer.Close)
 
 	return dynamicChallengeServer
 }
 
-// utf16FromString returns the UTF-16 encoding of the UTF-8 string s, with a terminating NUL added.
-// If s contains a NUL byte at any location, it returns (nil, syscall.EINVAL).
+// utf16FromString returns the UTF-16 encoding of the UTF-8 string s, with a
+// terminating NUL added. If s contains a NUL byte at any location, it
+// returns (nil, syscall.EINVAL).
 func utf16FromString(s string) ([]uint16, error) {
 	for i := 0; i < len(s); i++ {
 		if s[i] == 0 {
