+++ ee/server/service/scep/sceptest/testdata/testca/ca.key
